+++ llrt_core/src/modules/mod.rs
@@ -5,7 +5,7 @@ pub use llrt_modules::console;
 pub use llrt_modules::{
     abort, assert, buffer, child_process, crypto, dns, events, exceptions, fs, http,
     module_builder, navigator, net, os, path, perf_hooks, process, stream_web, string_decoder,
-    timers, tty, url, util, zlib,
+    timers, tty, url, util, zlib, ModuleNames,
 };
 
 #[cfg(feature = "lambda")]
